metric counting is causing problems on 32bit systems.

+++ src/noit_check.h
@@ -188,7 +188,7 @@ typedef struct noit_check {
 API_EXPORT(void) noit_poller_init();
 API_EXPORT(u_int64_t) noit_check_completion_count();
 API_EXPORT(u_int64_t) noit_check_metric_count();
-API_EXPORT(void) noit_check_metric_count_add(u_int64_t);
+API_EXPORT(void) noit_check_metric_count_add(int);
 API_EXPORT(int) noit_poller_check_count();
 API_EXPORT(int) noit_poller_transient_check_count();
 API_EXPORT(void) noit_poller_reload(const char *xpath); /* NULL for all */
